Shop - Realization
Completely rewrote the implementation of Balance - Garbage.

Added Ui to the shop.
Ability to buy things.
Adding items to the inventory.

I slightly improved the code TryAddItem - added two boolean variables. 1) Responsible for whether it is necessary to check whether objects are stacked or not. The 2nd one is responsible for checking whether it is necessary to add an object.